requirements.txt
pdfplumber>=0.11.0
Flask>=2.3.0
Werkzeug>=2.3.0
anthropic>=0.39.0
python-dotenv>=1.0.0
